+++ Classes/CrawlStrategy/GuzzleExecutionStrategy.php
@@ -38,7 +38,7 @@ class GuzzleExecutionStrategy implements LoggerAwareInterface, CrawlStrategy
     /**
      * Sets up a CURL / Guzzle Request for fetching the request.
      *
-     * @return bool|mixed
+     * @return array|bool|mixed
      */
     public function fetchUrlContents(UriInterface $url, string $crawlerId)
     {
@@ -62,7 +62,9 @@ class GuzzleExecutionStrategy implements LoggerAwareInterface, CrawlStrategy
                 sprintf('Error while opening "%s" - ' . $message, $url),
                 ['crawlerId' => $crawlerId]
             );
-            return $message;
+            return [
+                'errorlog' => [$message],
+            ];
         } catch (ConnectException $e) {
             $message = $e->getCode() . chr(32) . $e->getMessage();
 
@@ -70,7 +72,9 @@ class GuzzleExecutionStrategy implements LoggerAwareInterface, CrawlStrategy
                 sprintf('Error while opening "%s" - ' . $message, $url),
                 ['crawlerId' => $crawlerId]
             );
-            return $message;
+            return [
+                'errorlog' => [$message],
+            ];
         }
     }